feat: add player service and correction in repository (Integer -> Long)

+++ Semana2/ligamx/src/main/java/com/esteban/ligamx/repository/PlayerRepository.java
@@ -3,5 +3,5 @@ package com.esteban.ligamx.repository;
 import com.esteban.ligamx.model.Player;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlayerRepository extends JpaRepository<Player, Integer> {
+public interface PlayerRepository extends JpaRepository<Player, Long> {
 }
